+++ apps/admin-ui-operations/src/views/operations/operations-list-panel.tsx
@@ -12,6 +12,7 @@ import { matchPath } from 'react-router';
 
 import { DONE_ROUTE_ID, QUEUED_ROUTE_ID, RUNNING_ROUTE_ID } from '../../constants';
 import { type ManageOption } from '../../types/operations';
+import { SECTION_ROUTES } from './operations-section-routes';
 
 const VALID_TABS = new Set([RUNNING_ROUTE_ID, QUEUED_ROUTE_ID, DONE_ROUTE_ID]);
 
@@ -26,23 +27,12 @@ const OperationsListPanel: FC = () => {
 	}, [relativePathname]);
 
 	const manageOptions = useMemo<Array<ManageOption>>(
-		() => [
-			{
-				id: RUNNING_ROUTE_ID,
-				name: t('label.running', 'Running'),
+		() =>
+			SECTION_ROUTES.map(({ id, labelKey, labelDefault }) => ({
+				id,
+				name: t(labelKey, labelDefault),
 				isSelected: true
-			},
-			{
-				id: QUEUED_ROUTE_ID,
-				name: t('label.queued', 'Queued'),
-				isSelected: true
-			},
-			{
-				id: DONE_ROUTE_ID,
-				name: t('label.done', 'Done'),
-				isSelected: true
-			}
-		],
+			})),
 		[t]
 	);