+++ packages/ui-shared/src/exports.ts
@@ -68,7 +68,6 @@ import { useBreakpoint } from './hooks/use-breakpoint';
 import { useDebouncedValue } from './hooks/use-debounced-value';
 import { useLocalStorage } from './hooks/use-local-storage';
 import { useMediaQuery } from './hooks/use-media-query';
-import { useTotalQuotaActive } from './hooks/use-total-quota-active';
 import {
   type CloseSnackbarFn,
   type CreateSnackbarFn,
@@ -134,9 +133,7 @@ import {
   type GetCoreAttributesResponse,
 } from './services/get-core-attributes';
 import { getDomainInformation } from './services/get-domain-information';
-import { type FileQuotaResponse, getFileQuotaById } from './services/get-file-quota';
 import { getAllNotifications, readUnreadNotification } from './services/notification-service';
-import { resetFileQuotaLimitById } from './services/reset-file-quota-limit';
 import { getCosList } from './services/search-cos-service';
 import {
   type DirectoryAttribute,
@@ -146,9 +143,8 @@ import {
   type SearchDomainDirectories,
 } from './services/search-directory-service';
 import { setCoreAttributes } from './services/set-core-attributes';
-import { setFileQuotaLimitById } from './services/set-file-quota-limit';
 import { useDetailViewMaxWidth, usePrimaryBarState } from './shell/hooks';
-import { getApp, getShell, useAppList, useAppRoutes } from './store/app/hooks';
+import { getApp, getShell, useAppList, useAppRoutes, useModuleCrumbMenu } from './store/app/hooks';
 import { useAppStore } from './store/app/store';
 import { normalizeRoute } from './store/app/utils';
 import { useBridge, useContextBridge } from './store/context-bridge';
@@ -236,7 +232,6 @@ export {
   getCosGeneralInformation,
   getCosList,
   getDomainInformation,
-  getFileQuotaById,
   getLocale,
   getResponsiveContainerStyle,
   getResponsiveMaxWidth,
@@ -261,13 +256,11 @@ export {
   registerApp,
   removeRoute,
   replaceHistory,
-  resetFileQuotaLimitById,
   SCALING_LIMIT,
   SCALING_OPTIONS,
   searchDirectory,
   SEND_FEEDBACK_URL,
   setCoreAttributes,
-  setFileQuotaLimitById,
   SHELL_APP_ID,
   SnackbarManagerContext,
   soapFetch,
@@ -303,6 +296,7 @@ export {
   useLoginConfigStore,
   useMailstoreServers,
   useMediaQuery,
+  useModuleCrumbMenu,
   useModuleLicenseInfo,
   useMtaServers,
   usePrimaryBarState,
@@ -311,7 +305,6 @@ export {
   useServerVersion,
   useSnackbar,
   useStickyBarStore,
-  useTotalQuotaActive,
   useUserAccount,
   useUserAccounts,
   useUserSettings,
@@ -330,7 +323,6 @@ export type {
   DirectoryAttribute,
   DirectoryEntry,
   DomainDirectories,
-  FileQuotaResponse,
   GetCoreAttributesResponse,
   GetCosResponse,
   SearchDomainDirectories,